pyproject.toml
[project]
name = "social-campaign-agent"
version = "0.1.0"
description = "Multi-agent creative automation pipeline for social ad campaigns"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4",
    "langchain-openai>=0.3",
    "openai>=1.0",
    "Pillow>=11.0",
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "Jinja2>=3.0",
    "python-dotenv>=1.0",
]

[project.scripts]
social-campaign = "social_campaign.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/social_campaign"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
